Create Poker Hand Probabilities
This is a personal project executed out of self interest in poker and machine learning. This object-class returns the probability of the next card being a specific suit or rank and also the probabilities of the next one or two cards completing a hand. It can achieve this with any number of inputs from the hole or community cards. The purpose of this project is to not only gain a better understanding of the game of poker, but also to use as input to a machine learning model to make an artificially intelligent opponent.